pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tabletcraft"
version = "0.2.0"
description = "Turn any text into cuneiform clay tablets - translate, transliterate, and render ancient Mesopotamian writing"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [{name = "Geoffrey Wang", email = "geoffreywang1117@gmail.com"}]
keywords = ["cuneiform", "akkadian", "translation", "ancient-languages", "NLP", "clay-tablet"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "torch>=2.0",
    "transformers>=4.35",
    "sacrebleu>=2.3",
    "pandas",
    "numpy",
    "Pillow>=9.0",
    "svgwrite>=1.4",
]

[project.optional-dependencies]
serve = ["gradio>=4.0", "fastapi", "uvicorn"]
dev = ["pytest", "black", "ruff"]

[project.scripts]
tabletcraft = "tabletcraft.interfaces.cli:main"

[project.urls]
Homepage = "https://github.com/geoffreywang1117/tabletcraft"
Documentation = "https://github.com/geoffreywang1117/tabletcraft#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["tabletcraft*"]
